refactor: move dashboard from /app to /frontend
